move AddTransaction to accountController
